Error correction in readme. Moves obstacle and goal to separate folder in examples.
