+++ README.md
@@ -12,7 +12,7 @@ see https://github.com/maxspahn/motion_planning_scenes
 You can either install that manually as explained in the corresponding readme or 
 you install it through
 ```bash
-pip3 install -r requirements_scenes.py
+pip3 install -r requirements_scenes.txt
 ```
 
 ## Installation
