+++ urdfGymExamples/pointRobot_example.py
@@ -20,14 +20,14 @@ def main():
         ob = env.reset(pos=pos0, vel=vel0)
         env.setWalls(limits=[[-3, -2], [3, 2]])
         if obstacles:
-            from urdfGymExamples.obstacles import sphereObst1, sphereObst2, urdfObst1, dynamicSphereObst1
+            from urdfGymExamples.sceneObjects.obstacles import sphereObst1, sphereObst2, urdfObst1, dynamicSphereObst1
 
             env.addObstacle(sphereObst1)
             env.addObstacle(sphereObst2)
             env.addObstacle(urdfObst1)
             env.addObstacle(dynamicSphereObst1)
         if goal:
-            from urdfGymExamples.goal import goal1
+            from urdfGymExamples.sceneObjects.goal import goal1
 
             env.addGoal(goal1)
         print("Starting episode")
