+++ urdfGymExamples/sceneObjects/goal.py
