+++ urdfGymExamples/sceneObjects/obstacleData/duck.obj
