+++ urdfGymExamples/sceneObjects/obstacleData/duck.urdf
