+++ urdfGymExamples/sceneObjects/obstacleData/duck_vhacd.obj
