+++ urdfGymExamples/sceneObjects/obstacleData/sphere_goal.urdf
