+++ urdfGymExamples/sceneObjects/obstacleData/sphere_smooth.obj
