+++ urdfGymExamples/sceneObjects/obstacles.py
